requirements.txt
dj-database-url==0.5.0
Django==2.2
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
gunicorn==19.9.0
Markdown==3.1
psycopg2==2.7.7
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
whitenoise==4.1.3
